Merge the 'commands' config rather than replacing it when loading multiple config files.
With this change, if you have a `~/.config/projectable/config.toml` file that contains...

```
[commands]
"ctrl-t" = "echo hey there"
"ctrl-y" = "echo this is a demo"
```

and then a `$PWD/.projectable.toml` file that contains...

```
[commands]
"ctrl-t" = "echo this is replaced"
"ctrl-u" = "echo this is another command"
```

Then when you run `prj` in that `$PWD` (or below), you'll get three commands:
one for `ctrl-t`, one for `ctrl-y`, and one for `ctrl-u`.
The command bound to `ctrl-t` will be the one that in this example says "replaced".

Previously, the whole command map was getting replaced,
which meant in this scenario you'd only end up with `ctrl-t` and `ctrl-u`
(the `ctrl-y` command from the parent config just got discarded).

I suppose this could be considered a breaking change, but for me,
it's what I expected when I first read the docs about the built-in
configuration merging, so hopefully it's not too spicy :)